refactor solver
