+++ include/NeoFOAM/linearAlgebra/utilities.hpp
@@ -1,38 +1,54 @@
 // SPDX-License-Identifier: MIT
-// SPDX-FileCopyrightText: 2024 NeoFOAM authors
+// SPDX-FileCopyrightText: 2024-2025 NeoFOAM authors
 
 #pragma once
 
-namespace NeoFOAM
+#include <ginkgo/ginkgo.hpp>
+#include <ginkgo/extensions/kokkos.hpp>
+
+#include "NeoFOAM/fields/field.hpp"
+#include "NeoFOAM/core/dictionary.hpp"
+#include "NeoFOAM/linearAlgebra/linearSystem.hpp"
+
+
+namespace NeoFOAM::la
 {
 
-struct Dim
+std::shared_ptr<gko::Executor> getGkoExecutor(Executor exec)
 {
-    size_t numRows;
-    size_t numCols;
-};
+    return std::visit(
+        [](auto concreteExec)
+        { return gko::ext::kokkos::create_executor(concreteExec.underlyingExec()); },
+        exec
+    );
+}
 
-
-template<typename IndexType>
-struct MatrixCoordinate
+namespace detail
 {
-    IndexType row;
-    IndexType col;
-};
-
 template<typename ValueType, typename IndexType>
-struct MatrixEntry
+std::shared_ptr<gko::matrix::Csr<ValueType, IndexType>>
+createGkoMtx(std::shared_ptr<const gko::Executor> exec, LinearSystem<ValueType, IndexType>& sys)
 {
-    IndexType row;
-    IndexType col;
-    ValueType value;
-};
+    auto& mtx = sys.matrix();
+    size_t nrows = sys.rhs().size();
 
+    auto valuesView = gko::array<scalar>::view(exec, mtx.values().size(), mtx.values().data());
+    auto colIdxView = gko::array<int>::view(exec, mtx.colIdxs().size(), mtx.colIdxs().data());
+    auto rowPtrView = gko::array<int>::view(exec, mtx.rowPtrs().size(), mtx.rowPtrs().data());
 
-struct CompatibleWithAnyExecutor
+    return gko::share(gko::matrix::Csr<scalar, int>::create(
+        exec, gko::dim<2> {nrows, nrows}, valuesView, colIdxView, rowPtrView
+    ));
+}
+
+template<typename ValueType>
+std::shared_ptr<gko::matrix::Dense<ValueType>>
+createGkoDense(std::shared_ptr<const gko::Executor> exec, ValueType* ptr, size_t size)
 {
-    static Executor getCompatibleExecutor(const Executor& exec) { return exec; }
-};
-
+    return gko::share(gko::matrix::Dense<ValueType>::create(
+        exec, gko::dim<2> {size, 1}, gko::array<scalar>::view(exec, size, ptr), 1
+    ));
+}
+}
 
 }
